Massive game-mode and utils overhaul, doc consolidation, tests
- New game-mode DSL with join/disconnect policies, late-join,
  overtime, reconnection, respawn, activity watchdog
- Bot AI rewrite split into combat/construction/crafting/gathering/
  inventory/movement/mining/survival/utility goals & actions,
  exploration memory, sequence actions
- New utils: anticheat checks, drain manager, audit, replay flushes
- Drop per-utility .md files in favor of consolidated UTILS.md
- Bug fixes across schematic, scoreboard, screen, statue, tablist,
  worldedit, vanilla integration
- Add tests for LootTable, BotPersonality, BlockPositionIndex,
  MapPool, Region, Knockback, Noise, ViolationTracker, Condition

+++ src/main/kotlin/me/nebula/orbit/utils/mapgen/CaveGenerator.kt
@@ -245,7 +245,7 @@ class CaveCarver(
         for (dx in -rInt..rInt) {
             for (dy in -rInt..rInt) {
                 for (dz in -rInt..rInt) {
-                    val dist = sqrt((dx * dx + dy * dy * 0.6 + dz * dz).toDouble())
+                    val dist = sqrt(dx * dx + dy * dy * 0.6 + dz * dz)
                     if (dist > radius) continue
                     val bx = x + dx
                     val by = y + dy
