+++ src/main/kotlin/me/nebula/orbit/utils/snapshot/InventorySnapshot.kt
@@ -37,7 +37,7 @@ data class InventorySnapshot(
                 food = player.food,
                 health = player.health,
                 experience = player.exp,
-                level = player.level.toInt(),
+                level = player.level,
             )
         }
     }
